Feature: s3 cli integration
This commits integrate the logic for the `scw s3` commands. This
is using the minio client, preconfigured for `nl-ams` and `fr-par`
regions.
The minio configuration directory is computed from the location
of the .scwrc, creating a .scw directory at the same location.
The case of two .scwrc in the same directory is manage by adding
another directory layer inside .scw/, for example
/home/toto/.scwrc-toto will create configuration files for minio
under /home/toto/.scw/scwrc-toto/