feat(secret-manager): add option 'field' and 'raw' to secret version access
Add a new option 'field' to the secret version access subcommand. This
option allows to retrieve only the JSON field with the given name from a
Secret Version. The Secret Version data must be in JSON format.

Add a new option 'raw' to the secret version access subcommand. This
option allows to retrieve only the raw payload of a Secret Version.