fix(instance): cannot create server with project-id
While creating a server with a specific project-id, the request fail at
the IP creation step.
This is because we provide both a project and an organization while only
one of these is expected. We end up with the following error:

  Error while creating your public IP: scaleway-sdk-go: invalid argument(s):  does not respect constraint, Must contain one of ['organization', 'project']

This patch only provides a project since the organization field is
deprecated.