feat(instance): allow to create IPv6 only instances
With the routed IP, when creating a new instances with `ip=<IPv6_UUID>`
you still end up with an IPv4 attached to your instance due to the
`dynamic_ip_required` defaulting to `true`.
This change allows one to create IPv6 only instances when passing the
new option `ipv4=false` (which defaults to `true` to remain backward
compatible).